Move agent_receiver to cmk.agent_receiver
