+++ packages/cmk-agent-receiver/cmk/agent_receiver/decompression.py
@@ -32,7 +32,7 @@ class Decompressor(Enum):
         >>> Decompressor._zlib_decompress(b"blablub")
         Traceback (most recent call last):
             ...
-        agent_receiver.decompression.DecompressionError: ...
+        cmk.agent_receiver.decompression.DecompressionError: ...
         """
         try:
             return decompress(data)
