+++ packages/cmk-agent-receiver/tests/test_endpoints.py
@@ -17,11 +17,11 @@ from fastapi.testclient import TestClient
 from pydantic import UUID4
 from pytest_mock import MockerFixture
 
-from agent_receiver import site_context
-from agent_receiver.certs import serialize_to_pem
-from agent_receiver.checkmk_rest_api import CMKEdition, HostConfiguration, RegisterResponse
-from agent_receiver.models import ConnectionMode, R4RStatus, RequestForRegistration
-from agent_receiver.utils import R4R
+from cmk.agent_receiver import site_context
+from cmk.agent_receiver.certs import serialize_to_pem
+from cmk.agent_receiver.checkmk_rest_api import CMKEdition, HostConfiguration, RegisterResponse
+from cmk.agent_receiver.models import ConnectionMode, R4RStatus, RequestForRegistration
+from cmk.agent_receiver.utils import R4R
 
 from .certs import generate_csr_pair
 
@@ -63,7 +63,7 @@ def test_register_existing_ok(
         return RegisterResponse(connection_mode=ConnectionMode.PULL)
 
     mocker.patch(
-        "agent_receiver.endpoints.register",
+        "cmk.agent_receiver.endpoints.register",
         rest_api_register_mock,
     )
     response = client.post(
@@ -121,7 +121,7 @@ def test_register_register_with_hostname_host_missing(
     uuid: UUID4,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.host_configuration",
+        "cmk.agent_receiver.endpoints.host_configuration",
         side_effect=HTTPException(
             status_code=404,
             detail="N O T  F O U N D",
@@ -145,7 +145,7 @@ def test_register_register_with_hostname_wrong_site(
     uuid: UUID4,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.host_configuration",
+        "cmk.agent_receiver.endpoints.host_configuration",
         return_value=HostConfiguration(
             site="some-site",
             is_cluster=False,
@@ -174,7 +174,7 @@ def test_register_register_with_hostname_cluster_host(
     uuid: UUID4,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.host_configuration",
+        "cmk.agent_receiver.endpoints.host_configuration",
         return_value=HostConfiguration(
             site="NO_SITE",
             is_cluster=True,
@@ -198,14 +198,14 @@ def test_register_register_with_hostname_unauthorized(
     uuid: UUID4,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.host_configuration",
+        "cmk.agent_receiver.endpoints.host_configuration",
         return_value=HostConfiguration(
             site="NO_SITE",
             is_cluster=False,
         ),
     )
     mocker.patch(
-        "agent_receiver.endpoints.link_host_with_uuid",
+        "cmk.agent_receiver.endpoints.link_host_with_uuid",
         side_effect=HTTPException(
             status_code=403,
             detail="You do not have the permission for agent pairing.",
@@ -247,13 +247,13 @@ def test_register_register_with_hostname_hostname_validity(
     valid: bool,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.host_configuration",
+        "cmk.agent_receiver.endpoints.host_configuration",
         return_value=HostConfiguration(
             site="NO_SITE",
             is_cluster=False,
         ),
     )
-    mocker.patch("agent_receiver.endpoints.link_host_with_uuid", return_value=None)
+    mocker.patch("cmk.agent_receiver.endpoints.link_host_with_uuid", return_value=None)
 
     response = client.post(
         "/register_with_hostname",
@@ -279,7 +279,7 @@ def test_register_new_unauthenticated(
     serialized_csr: str,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         side_effect=HTTPException(
             status_code=401,
             detail="User authentication failed",
@@ -305,7 +305,7 @@ def test_register_new_cre(
     serialized_csr: str,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         return_value=CMKEdition.cre,
     )
     response = client.post(
@@ -329,7 +329,7 @@ def test_register_new_uuid_csr_mismatch(
     serialized_csr: str,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         return_value=CMKEdition.cce,
     )
     response = client.post(
@@ -352,7 +352,7 @@ def _test_register_new(
     serialized_csr: str,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         return_value=CMKEdition.cce,
     )
     response = client.post(
@@ -419,7 +419,7 @@ def test_register_new_ongoing_cre(
     uuid: UUID4,
 ) -> None:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         return_value=CMKEdition.cre,
     )
     response = client.post(
@@ -438,7 +438,7 @@ def _call_register_new_ongoing_cce(
     uuid: UUID4,
 ) -> httpx.Response:
     mocker.patch(
-        "agent_receiver.endpoints.cmk_edition",
+        "cmk.agent_receiver.endpoints.cmk_edition",
         return_value=CMKEdition.cce,
     )
     return client.post(
